+++ forge/changelog.md
@@ -1,2 +1,2 @@
-- Updated korean translation
-- Fixed LAN server message appearing twice
+- Updated to newest Forge version
+- Updated render event names